+++ animal-creation-starter/src/main/java/ru/mts/hw/config/CreateAnimalBeanPostProcessor.java
@@ -1,0 +1,23 @@
+package ru.mts.hw.config;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanPostProcessor;
+import org.springframework.lang.Nullable;
+import ru.mts.hw.service.CreateAnimalService;
+
+public class CreateAnimalBeanPostProcessor implements BeanPostProcessor {
+    @Nullable
+    @Override
+    public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
+        return bean;
+    }
+
+    @Nullable
+    @Override
+    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
+        if (bean instanceof CreateAnimalService) {
+            ((CreateAnimalService) bean).initAnimal();
+        }
+        return bean;
+    }
+}
